backend/django/requirements.txt
asgiref==3.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
filelock==3.20.0
flake8==7.1.1
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.10
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vaderSentiment==3.3.2
